move method fit_sgd from BaseHMM to SSM so that this can be used directly for lgssm model (tested); remove MAP estimation for the moment from lgssm since currently it's not jitted; revert the posterior update function to previous version that does not contain integer conversion
