+++ ssm_jax/hmm/models/base.py
@@ -181,51 +181,6 @@ class BaseHMM(SSM):
         self.unconstrained_params = params
         return jnp.array(log_probs)
 
-    def fit_sgd(self,
-                batch_emissions,
-                optimizer=optax.adam(1e-3),
-                batch_size=1,
-                num_epochs=50,
-                shuffle=False,
-                key=jr.PRNGKey(0),
-                **batch_covariates):
-        """
-        Fit this HMM by running SGD on the marginal log likelihood.
-        Note that batch_emissions is initially of shape (N,T)
-        where N is the number of independent sequences and
-        T is the length of a sequence. Then, a random susbet with shape (B, T)
-        of entire sequence, not time steps, is sampled at each step where B is
-        batch size.
-        Args:
-            batch_emissions (chex.Array): Independent sequences.
-            optmizer (optax.Optimizer): Optimizer.
-            batch_size (int): Number of sequences used at each update step.
-            num_epochs (int): Iterations made through entire dataset.
-            shuffle (bool): Indicates whether to shuffle minibatches.
-            key (chex.PRNGKey): RNG key to shuffle minibatches.
-        Returns:
-            losses: Output of loss_fn stored at each step.
-        """
-        def _loss_fn(params, minibatch_emissions, **minibatch_covariates):
-            """Default objective function."""
-            self.unconstrained_params = params
-            scale = len(batch_emissions) / len(minibatch_emissions)
-            minibatch_lls = vmap(self.marginal_log_prob)(minibatch_emissions, **minibatch_covariates)
-            lp = self.log_prior() + minibatch_lls.sum() * scale
-            return -lp / batch_emissions.size
-
-        params, losses = run_sgd(_loss_fn,
-                                 self.unconstrained_params,
-                                 batch_emissions,
-                                 optimizer=optimizer,
-                                 batch_size=batch_size,
-                                 num_epochs=num_epochs,
-                                 shuffle=shuffle,
-                                 key=key,
-                                 **batch_covariates)
-        self.unconstrained_params = params
-        return losses
-
 
 class StandardHMM(BaseHMM):
